backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.3
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.1
pydantic[email]==2.6.4
